Use viminfo/ShaDa instead of mkview/loadview
Remove the autocmd for mkview and loadview, and use the autocmd for
restoring last cursor position from viminfo/ShaDa (:help restore-cursor).
This does mean that I lose saving of folds state, but it is a much
simpler solution. If I miss the fold state restoring functionality I can
revert it from here.
Also simplify directories settings for swap and undofile. Not using
backup so removed that completely. Viminfo/ShaDa set to default for now.